prevent the false setting of glance
After the execution of folsom test, '/opt/stack/glance/bin' is left.
devstack decides a path of glance-api by the existence of the directory.
glance of grizzly does not install glance-api to the directory.
Therefore, in grizzly test just after folsom, start of g-api fails.

This change remove the directory before start of devstack to prevent the
failure.