update test script
This patch removes the check of the validity of the dhcp server
which is performed before creating VM.
The dhcp server is started when VM port is created.
This patch also changes the test for overlapping IP.
launch VM with a specific IP address instead of checking an IP
address of VM after VM was launched.